Added some code that will be usefull for foundation testing that we found useful in vbs
