+++ MK6.AutomatedTesting.UI/MK6.AutomatedTesting.UI.csproj
@@ -123,8 +123,9 @@
     <Compile Include="Configuration\DriverConfigurationSection.cs" />
     <Compile Include="DataTableLoader.cs" />
     <Compile Include="DriverFactory.cs" />
+    <Compile Include="Extensions\PageExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="WebDriverExtensions.cs" />
+    <Compile Include="Extensions\WebDriverExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
